don't refreshify event handlers outside render cycle
