+++ package.json
@@ -23,6 +23,7 @@
     "mocha": "^2.1.0",
     "trytryagain": "^0.1.0",
     "uglify-js": "^2.4.16",
+    "virtual-dom-stringify": "^2.1.0",
     "watchify": "^2.2.1"
   },
   "scripts": {
